+++ lib/src/extensions.dart
@@ -4,6 +4,8 @@
 
 import 'package:flutter/material.dart';
 
+import '../calendar_view.dart';
+
 extension DateTimeExtensions on DateTime {
   /// Compares only [day], [month] and [year] of [DateTime].
   bool compareWithoutTime(DateTime date) {
@@ -34,7 +36,8 @@ extension DateTimeExtensions on DateTime {
 
   /// Gets difference of weeks between [date] and calling object.
   int getWeekDifference(DateTime date) =>
-      (firstDayOfWeek.difference(date.firstDayOfWeek).inDays.abs() / 7).ceil();
+      (firstDayOfWeek().difference(date.firstDayOfWeek()).inDays.abs() / 7)
+          .ceil();
 
   /// Returns The List of date of Current Week
   /// Day will start from Monday to Sunday.
@@ -43,37 +46,45 @@ extension DateTimeExtensions on DateTime {
   /// will return dates
   /// [6,7,8,9,10,11,12]
   /// Where on 6th there will be monday and on 12th there will be Sunday
-  List<DateTime> datesOfWeek() {
-    // weekday can very from 1(Monday) to 7(Sunday).
-    // If we have a date that has week day of 3 (Wednesday) and
-    // we want to find first date of the week that contains this date
-    // we need to move 2 days back. Wednesday -> Tuesday -> Monday.
-    // That is same as 3 - 1 (weekday - 1).
+  List<DateTime> datesOfWeek({WeekDays start = WeekDays.monday}) {
+    // Here %7 ensure that we do not subtract >6 and <0 days.
+    // Initial formula is,
+    //    difference = (weekday - startInt)%7
+    // where weekday and startInt ranges from 1 to 7.
+    // But in WeekDays enum index ranges from 0 to 6 so we are
+    // adding 1 in index. So, new formula with WeekDays is,
+    //    difference = (weekdays - (start.index + 1))%7
     //
-    final start = subtract(Duration(days: weekday - 1));
+    final startDay = subtract(Duration(days: (weekday - start.index - 1) % 7));
 
     return [
-      start,
-      start.add(Duration(days: 1)),
-      start.add(Duration(days: 2)),
-      start.add(Duration(days: 3)),
-      start.add(Duration(days: 4)),
-      start.add(Duration(days: 5)),
-      start.add(Duration(days: 6)),
+      startDay,
+      startDay.add(Duration(days: 1)),
+      startDay.add(Duration(days: 2)),
+      startDay.add(Duration(days: 3)),
+      startDay.add(Duration(days: 4)),
+      startDay.add(Duration(days: 5)),
+      startDay.add(Duration(days: 6)),
     ];
   }
 
-  DateTime get firstDayOfWeek => subtract(Duration(days: weekday - 1));
-  DateTime get lastDayOfWeek => add(Duration(days: 7 - weekday));
+  /// Returns the first date of week containing the current date
+  DateTime firstDayOfWeek({WeekDays start = WeekDays.monday}) =>
+      subtract(Duration(days: (weekday - start.index - 1) % 7));
+
+  /// Returns the last date of week containing the current date
+  DateTime lastDayOfWeek({WeekDays start = WeekDays.monday}) =>
+      add(Duration(days: 6 - (weekday - start.index - 1) % 7));
 
   /// Returns list of all dates of [month].
   /// All the dates are week based that means it will return array of size 42
   /// which will contain 6 weeks that is the maximum number of weeks a month
   /// can have.
-  List<DateTime> get datesOfMonths {
+  List<DateTime> datesOfMonths({WeekDays startDay = WeekDays.monday}) {
     final monthDays = <DateTime>[];
     for (var i = 1, start = 1; i < 7; i++, start += 7) {
-      monthDays.addAll(DateTime(year, month, start).datesOfWeek());
+      monthDays
+          .addAll(DateTime(year, month, start).datesOfWeek(start: startDay));
     }
     return monthDays;
   }